requirements.txt
django
django-daterange-filter
django-inline-actions
-e git://github.com/digi604/django-smart-selects#egg=django-smart-selects
django-autoslug
